requirements.txt
tensorflow-cpu
numpy
Pillow
streamlit
fastapi
uvicorn
python-multipart
